Fixes validation to work for floats

+++ src/data/utils/validators.py
@@ -41,8 +41,9 @@ def validate_timestamp(ts):
         return True
     pass
 
+#!! NEEDS VALIDATION FOR (interval, adj_open, adj_high, adj_low, adj_close, adj_volume, div_cash, split_factor) !!
 def validate_candle(open_price, high, low, close, volume):
-    if type(open_price) != int or type(high) != int or type(low) != int or type(close) != int or type(volume) != int:
+    if type(open_price) != float or type(high) != float or type(low) != float or type(close) != float or type(volume) != int:
         return False
     if open_price < 0 or high < 0 or low < 0 or volume < 0 or close < 0: #checks if any of the variables are negative
         return False
